legacy_pygame/requirements.txt
contourpy==1.3.3
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
et_xmlfile==2.0.0
fonttools==4.61.1
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.1.0
pycparser==2.23
pygame-ce==2.5.6
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
seaborn==0.13.2
six==1.17.0
tinycss2==1.5.1
tzdata==2025.3
webencodings==0.5.1
